# lib/protocol: Optimize FileKey