Use lib/sha256 to get the fast Minio hash, allocate the salt at init time. Benchmark results on Linux/amd64 with STHASHING=minio:

```
name       old time/op    new time/op    delta
FileKey-8    6.27µs ± 3%    4.58µs ± 3%  -26.91%  (p=0.000 n=20+19)

name       old alloc/op   new alloc/op   delta
FileKey-8    1.47kB ± 0%    1.17kB ± 0%  -20.58%  (p=0.000 n=20+20)

name       old allocs/op  new allocs/op  delta
FileKey-8      19.0 ± 0%      16.0 ± 0%  -15.79%  (p=0.000 n=20+20)
```